Moved components and updated imports
