+++ components/ProductModal/index.ts
@@ -1,0 +1,3 @@
+import ProductModal from "./ProductModal";
+
+export default ProductModal;
