+++ components/types/product.ts
@@ -1,0 +1,7 @@
+import { Addon, AddonGroup, Product } from "@prisma/client";
+
+export type ProductWithAddons = Product & {
+  addonGroups: (AddonGroup & {
+    addons: Addon[];
+  })[];
+};
